chore: ruff gate, dead code and comment purge, --wav sink fix
Ruff is now a dev dependency with a project rule set (scripts/check.ps1 runs
ruff then pytest). The 101 pre-existing findings are resolved or deliberately
ignored per rule; the four undefined names are fixed via TYPE_CHECKING
imports. The --wav path's sink took two arguments while the pipeline passes
three (a TypeError on every file transcription) — fixed and covered by
tests/test_cli.py. Unused constants, methods and two never-read
TranscriptSegment fields are removed. Comments are stripped down to the
handful that name a Windows, library or privacy constraint the code cannot
express; the cleanup buffer row becomes a dataclass and the enrollment phase
a Literal so their shape no longer lives in a comment.

+++ src/oat_notes/capture.py
@@ -87,7 +87,7 @@ class AudioCapture:
 
 
 def _resample(samples: np.ndarray, from_rate: int, to_rate: int) -> np.ndarray:
-    target_length = int(round(samples.size * to_rate / from_rate))
+    target_length = round(samples.size * to_rate / from_rate)
     positions = np.linspace(0, samples.size - 1, target_length)
     return np.interp(positions, np.arange(samples.size), samples).astype(np.float32)
